Report renaming.
1. Freaking Years Report -> Events Rangered Report
2. Service Years Report - Freaking Years report.
