+++ app/components/navbar-area-reports.hbs
@@ -46,8 +46,8 @@
     <NavbarMenuSection @section="Quartermaster">
       <NavbarMenuItem @menu={{menu}} @route="reports.alpha-shirts" @item="Alpha Shirts"/>
       {{#if (has-role "admin" "quartermaster")}}
+        <NavbarMenuItem @menu={{menu}} @route="reports.events-rangered" @item="Events Rangered"/>
         <NavbarMenuItem @menu={{menu}} @route="reports.freaking-years" @item="Freaking Years"/>
-        <NavbarMenuItem @menu={{menu}} @route="reports.service-years" @item="Service Years"/>
         <NavbarMenuItem @menu={{menu}} @route="reports.potential-shirts" @item="Potential Earned Shirts"/>
         <NavbarMenuItem @menu={{menu}} @route="reports.potential-swag" @item="Potential Swag"/>
         <NavbarMenuItem @menu={{menu}} @route="reports.swag-distribution" @item="Swag Distribution"/>
