+++ app/router.js
@@ -218,7 +218,7 @@ Router.map(function () {
     this.route('event-stats');
     this.route('flakes');
     this.route('forced-signins');
-    this.route('freaking-years');
+    this.route('events-rangered');
     this.route('hq-forecast');
     this.route('languages');
     this.route('no-shows');
@@ -233,7 +233,7 @@ Router.map(function () {
     this.route('sandman-qualified');
     this.route('schedule-by-callsign');
     this.route('schedule-by-position');
-    this.route('service-years');
+    this.route('freaking-years');
     this.route('shift-coverage');
     this.route('shift-drop');
     this.route('shift-lead');
