+++ app/templates/reports/freaking-years.hbs
@@ -1,24 +1,20 @@
 <h1>Freaking Years Report</h1>
 
 <p>
-  <RadioGroup @options={{this.showOptions}} @onChange={{set-value this 'showAll'}} @value={{this.showAll}} />
-</p>
-
-<p>
-  Report on people's service based on years rangered. The "Signed Up" column
-  indicates if ANY shift sign up (training, dirt shift, etc) was found for the
-  current year which may indicate the person will be working the event.
+  This report shows each person's total service years. Service years are calculated from a combination of events
+  rangered, Echelon contributions, and year-round recognitions — not solely from events rangered.
 </p>
 
 <p>
   <UiExportToCSVButton @onClick={{this.exportToCsv}} />
 </p>
 
-Showing {{this.totalPeople}} people over {{this.freaking.length}} years.
-{{#each this.freaking as |freaks|}}
+Showing {{this.totalPeople}} people over {{this.serviceYears.length}} years.
+{{#each this.serviceYears as |serviceYear|}}
   <div class="max-width-700">
     <UiAccordion as |accordion|>
-      <accordion.title>{{pluralize freaks.years "year"}} - {{pluralize freaks.people.length "person"}}</accordion.title>
+      <accordion.title>{{pluralize serviceYear.years "year"}} - {{pluralize serviceYear.people.length
+                                                                            "person"}}</accordion.title>
       <accordion.body>
         <UiTable>
           <thead>
@@ -28,11 +24,10 @@ Showing {{this.totalPeople}} people over {{this.freaking.length}} years.
             <th class="w-10">Status</th>
             <th>First Year</th>
             <th>Last Year</th>
-            <th>Signed Up {{this.signed_up_year}}</th>
           </tr>
           </thead>
           <tbody>
-          {{#each freaks.people key="id" as |person| }}
+          {{#each serviceYear.people key="id" as |person| }}
             <tr>
               <td class="w-20">
                 <PersonLink @person={{person}} />
@@ -41,7 +36,6 @@ Showing {{this.totalPeople}} people over {{this.freaking.length}} years.
               <td class="w-10">{{person.status}}</td>
               <td>{{person.first_year}}</td>
               <td>{{person.last_year}}</td>
-              <td class="text-center">{{yesno person.signed_up}}</td>
             </tr>
           {{/each}}
           </tbody>
