--- tests/unit/controllers/reports/freaking-years-test.js
@@ -1,12 +1,0 @@
-import { module, test } from 'qunit';
-import { setupTest } from 'ember-qunit';
-
-module('Unit | Controller | reports/freaking-years', function(hooks) {
-  setupTest(hooks);
-
-  // Replace this with your real tests.
-  test('it exists', function(assert) {
-    let controller = this.owner.lookup('controller:reports/freaking-years');
-    assert.ok(controller);
-  });
-});
