--- tests/unit/routes/reports/freaking-years-test.js
@@ -1,11 +1,0 @@
-import { module, test } from 'qunit';
-import { setupTest } from 'ember-qunit';
-
-module('Unit | Route | reports/freaking-years', function(hooks) {
-  setupTest(hooks);
-
-  test('it exists', function(assert) {
-    let route = this.owner.lookup('route:reports/freaking-years');
-    assert.ok(route);
-  });
-});
